Updated to version 3.0.0   * Adjusted .gitignore file   * Segregated mocha tests   * Added formatting to tests and fixed them all   * Updated WireMock standalone version
