+++ .mocharc.js
@@ -1,0 +1,16 @@
+'use strict';
+
+// This is a JavaScript-based config file containing every Mocha option plus others.
+// If you need conditional logic, you might want to use this type of config,
+// e.g. set options via environment variables 'process.env'.
+// Otherwise, JSON or YAML is recommended.
+
+module.exports = {
+    'async-only': true,
+    bail: false,
+    'check-leaks': false,
+    color: true,
+    'inline-diffs': true,
+    diff: true,
+    spec: ['test/**/*.js']
+}
